// src/generated/schema.ts
export interface components {
    schemas: {
// ...
        LoginResponseModel: {
            accessToken?: string;
            refreshToken?: string;
            expiresIn?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}